fix: ai 피드백 수정

+++ src/main/java/kr/muroom/muroombackendbach/user/domain/repository/MyStudioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MyStudioRepository extends JpaRepository<MyStudio, Integer> {
+public interface MyStudioRepository extends JpaRepository<MyStudio, Long> {
 
 }
